xclip/requirements.txt
torch
torchvision
torchaudio
torchmetrics
tensorflow
tensorflow_hub
spacy
wikipedia
pandas
tqdm
nltk
pillow
mmcv-full==1.7.0 
-f https://download.openmmlab.com/mmcv/dist/cu117/torch1.13/index.html
tables
wikipedia
transformers
datasets
sentence_transformers
matplotlib

terminaltables
pycocotools
git+https://github.com/openai/CLIP.git
git+https://github.com/modestyachts/ImageNetV2_pytorch

img2pdf
tiktoken
open_clip_torch